chore(check-tree-select): 代码优化
